Extract list helper into accounts utils file
